requirements.txt
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
distlib==0.3.6
filelock==3.8.0
idna==3.4
importlib-metadata==5.0.0
loguru==0.6.0
NetfilterQueue==1.0.0
paramiko==2.12.0
platformdirs==2.5.2
psutil==5.9.3
pyaml==21.10.1
pycparser==2.21
pyinotify==0.9.6
PyNaCl==1.5.0
PyYAML==6.0
requests==2.28.1
scapy==2.5.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.6
zipp==3.10.0
